// apps/api/src/types/kyc.ts
// src/kyc.ts
export enum KycStatus {
  PENDING = 'pending',
  SUCCESS = 'success',
  FAILED = 'failed',
}

export interface Kyc {
  id: number;
  customer_id: number;
  first_name?: string;
  last_name: string;
  location: string;
  birthdate: string;
  emergency_number_1: string;
  emergency_number_1_name: string;
  emergency_number_2?: string;
  emergency_number_2_name?: string;
  front_photo: string;
  back_photo: string;
  selfie: string;
  status: KycStatus;
  created_at: string; // Auto-generated by Supabase
  updated_at: string; // Auto-generated by Supabase
}

// Types pour les opérations CRUD
export type CreateKycInput = Omit<Kyc, 'id' | 'created_at' | 'updated_at'>;
export type UpdateKycInput = Partial<
  Omit<Kyc, 'id' | 'created_at' | 'customer_id'>
>;